+++ vim/hooks.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-source ../pottyfunctions.sh
-
 VIM_CONFIG_DIR=$HOME/.vim
 VIM_PLUGIN_DIR=$VIM_CONFIG_DIR/autoload
 
@@ -15,12 +13,22 @@ fi
 # Deploy vimrc in home directory
 deploy_dotfile "$PWD/dotfiles/vimrc" "$HOME/.vimrc"
 
-# Get latest version of vim plug
-curl -fLo ./plugins/plug.vim \
-	https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+if [[ "$QUICK_DEPLOY" == false ]]; then
+    # Get latest version of vim plug
+    curl -fLo ./plugins/plug.vim \
+            https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 
-# Place vim plug in vim's autoload plugin directory
-cp ./plugins/plug.vim $VIM_PLUGIN_DIR
+    # Place vim plug in vim plugin directory if it changed
+    if [[ -e $VIM_PLUGIN_DIR/plug.vim ]] && \
+            ! diff -q ./plugins/plug.vim $VIM_PLUGIN_DIR/plug.vim; then
+        cp ./plugins/plug.vim $VIM_PLUGIN_DIR
+    fi
+fi
+
+# Place vim plug in vim's autoload plugin directory if it doesn't exist
+if [[ ! -e $VIM_PLUGIN_DIR/plug.vim ]]; then
+    cp ./plugins/plug.vim $VIM_PLUGIN_DIR
+fi
 
 # Symlinks for neovim
 if [[ ! -e ~/.config/nvim ]]; then
